lambdas/process_page_batch/requirements.txt
# Core dependencies for the Lambda function
msgpack>=1.0.0

# Runtime dependencies
tabulate>=0.9.0,<1.0.0
pillow>=10.0.0,<12.0.0
pydantic>=2.0.0
jsonschema>=4.16.0,<5.0.0

# Minimal docling-core runtime deps for page types used in Lambda
typing-extensions<5.0.0,>=4.12.2
